updated stock table component
