+++ src/components/charts/StockPieChart.tsx
