+++ src/components/charts/index.ts
@@ -1,0 +1,7 @@
+const CHART_TYPES = [
+    "StockTable",
+    "StockPieChart",
+    "StockTreeMap"
+] as const
+
+export type ChartType = typeof CHART_TYPES[number]
